requirements.txt
asgiref==3.3.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.4
django-filter==2.4.0
django-import-export==2.5.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
importlib-metadata==4.5.0
Markdown==3.3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
pkg-resources==0.0.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
typing-extensions==3.10.0.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1
